# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KKN Candles API",
        "description": "API pour la boutique de bougies KKN Candles",
        "contact": {
            "name": "Support KKN Candles",
            "email": "support@kkncandles.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Détail d'une catégorie",
                "description": "Récupérer les détails d'une catégorie par ID ou slug",
                "operationId": "getCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ou slug de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            }
        },
        "/categories/{id}/products": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Produits d'une catégorie",
                "description": "Récupérer les produits d'une catégorie spécifique",
                "operationId": "getCategoryProducts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ou slug de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre de produits par page (max 50)",
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produits de la catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "category": {
                                                    "$ref": "#/components/schemas/Category"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Product"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            }
        },
        "/contact": {
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Envoyer un message de contact",
                "description": "Envoyer un message via le formulaire de contact",
                "operationId": "sendContactMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "subject",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Jean Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "jean@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+225 07 00 00 00 00",
                                        "nullable": true
                                    },
                                    "subject": {
                                        "type": "string",
                                        "maxLength": 200,
                                        "example": "Question sur une commande"
                                    },
                                    "message": {
                                        "type": "string",
                                        "maxLength": 2000,
                                        "example": "Bonjour, je souhaite avoir des informations sur..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Votre message a été envoyé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Modèle produit",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "name": {
                        "type": "string",
                        "example": "Bougie Parfumée Vanille"
                    },
                    "slug": {
                        "type": "string",
                        "example": "bougie-parfumee-vanille"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "long_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "integer",
                        "example": 5000
                    },
                    "formatted_price": {
                        "type": "string",
                        "example": "5 000 FCFA"
                    },
                    "stock_quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "url"
                        }
                    },
                    "main_image": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    },
                    "specifications": {
                        "type": "object",
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Modèle catégorie",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Bougies Parfumées"
                    },
                    "slug": {
                        "type": "string",
                        "example": "bougies-parfumees"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    },
                    "products_count": {
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Métadonnées de pagination",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 12
                    },
                    "total": {
                        "type": "integer",
                        "example": 120
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification JWT Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}